pyproject.toml
[tool.poetry]
name = "PDF ingest"
version = "1.0.0"
description = "Demonstrate the workflow to ingest PDF files, enrich images and create text markdowns"
authors = ["Tiran Dagan <tiran@tirandagan.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
pillow = "^11.0.0"
pymupdf = "^1.24.12"
matplotlib = "^3.9.2"
langchain-unstructured = "^0.1.5"
unstructured-ingest = "^0.1.1"
inquirer = "^3.4.0"
openai = "^1.52.1"
tabulate = "^0.9.0"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
